### 2018-09-24 #975 @arfon
@edawson - I think you're missing a comma on this line: https://github.com/edawson/gfakluge/blob/master/paper/paper.bib#L7
