### 2018-09-24 d4455c3
Fix a missing comma in paper.bib
