+++ paper/paper.bib
@@ -4,7 +4,7 @@
   year = 2018,
   url = {https://github.com/edawson/gfakluge}
 }
-@article{TwoPaCo
+@article{TwoPaCo,
     author = {Ilia Minkin, Son Pham, and Paul Medvedev},
     title = {TwoPaCo: An efficient algorithm to build the compacted de Bruijn graph from many complete genomes},
     journal = {Bioinformatics},
