### 2018-09-24 #975 @edawson
Thanks @arfon ! I've fixed the missing comma on the master branch - do I need to do anything else?
