+++ paper/paper.bib
@@ -32,9 +32,9 @@
     doi = {doi:10.1038/nbt.4227}
 }
 @article{Bandage,
-    author = {Wick R.R., Schultz M.B., Zobel J. and Holt K.E.}
-    year = {2015}
-    title = {Bandage: interactive visualisation of de novo genome assemblies}
+    author = {Wick R.R., Schultz M.B., Zobel J. and Holt K.E.},
+    year = {2015},
+    title = {Bandage: interactive visualisation of de novo genome assemblies},
     journal = {Bioinformatics},
     volume = {31},
     pages = {3350-3352},
