### 2018-10-04 c53a7bc
Clean up exit codes and add trim command functions to main.
