+++ src/main.cpp
@@ -29,9 +29,19 @@ int gfakluge_package_help(char** argv){
         "   sort:    Print a GFA file in HSLP / HSEFGUO order." << endl <<
         "   stats:   Get assembly statistics (e.g. N50) for a GFA file." << endl <<
         "   subset:  Extract the subgraph between two IDs in a graph." << endl <<
+        "   trim:    Remove elements from a GFA graph." << endl <<
         endl;
     return 1;
 
+}
+
+void trim_help(char** argv){
+    cerr << argv[0] <<  " trim: remove elements from a GFA graph." << endl
+        << "Usage: " << argv[0] << " trim [OPTIONS] <GFA_FILE> " << endl
+            << " -l / --length  <INT>  Remove segments (and their edges) if their sequence length is less than <INT>." << endl
+            << " -N / --no-ambiguous   Remove segments which have ambiguous bases (i.e. non-ATGC) in their sequence." << endl
+            << " -v / --version        print GFAK version and exit." << endl
+        << endl;
 }
 
 
@@ -142,6 +152,23 @@ void subset_help(char** argv){
 }
 
 /**
+ * Trim segments (and their edges) from a graph
+ */
+int trim_main(int argc, char** argv){
+    string gfa_gile;
+
+    if (argc < 3){
+        cerr << "No GFA file given as input." << endl << endl;
+        trim_help(argv);
+        exit(1);
+    }
+
+
+
+    return 0;
+}
+
+/**
  * Output a fasta file from input GFA
  */
 int extract_main(int argc, char** argv){
@@ -151,7 +178,7 @@ int extract_main(int argc, char** argv){
     if (argc < 3){
         cerr << "No GFA file given as input." << endl << endl; 
         extract_help(argv);
-        exit(0);
+        exit(1);
     }
 
     int c;
@@ -178,7 +205,7 @@ int extract_main(int argc, char** argv){
                 exit(0);
             case 'v':
                 print_version_help();
-                exit(1);
+                exit(0);
             case 'p':
                 include_paths= true;
                 break;
@@ -289,7 +316,7 @@ int fillseq_main(int argc, char** argv){
     if (argc < 3){
         cerr << "fillseq requires a GFA file." << endl << endl;
         fillseq_help(argv);
-        exit(9);
+        exit(1);
     }
 
     string fasta_file;
@@ -327,7 +354,7 @@ int fillseq_main(int argc, char** argv){
                 break;
             case 'v':
                 print_version_help();
-                exit(1);
+                exit(0);
             case 'S':
                 spec_version = stod(optarg);
                 break;
@@ -366,7 +393,7 @@ int fillseq_main(int argc, char** argv){
     else if (spec_version != 0.0){
         cerr << "Invalid specification number: " << spec_version << endl
         << "Please provide one of [0.1, 1.0, 2.0]." << endl;
-        exit(22);
+        exit(1);
     }
     
 
@@ -392,7 +419,7 @@ int diff_main(int argc, char** argv){
     if (argc < 4){
         cerr << "diff requires two GFA files as input." << endl << endl;
         diff_help(argv);
-        exit(0);
+        exit(1);
     }
 
     int c;
@@ -418,7 +445,7 @@ int diff_main(int argc, char** argv){
                 exit(0);
             case 'v':
                 print_version_help();
-                exit(1);
+                exit(0);
             default:
                 abort();
         }
@@ -472,7 +499,7 @@ int convert_main(int argc, char** argv){
     if (argc < 3){
         cerr << "No GFA file provided. Please provide a GFA file to convert" << endl;
         convert_help(argv);
-        exit(0);
+        exit(1);
     }
 
     int c;
@@ -500,7 +527,7 @@ int convert_main(int argc, char** argv){
             case '?':
             case 'v':
[…]
