### 2018-10-04 cc91753
Add functions to trim sequence_elems
