+++ src/main.cpp
@@ -155,15 +155,70 @@ void subset_help(char** argv){
  * Trim segments (and their edges) from a graph
  */
 int trim_main(int argc, char** argv){
-    string gfa_gile;
+    string gfa_file;
+
+    int minlen = 0;
+    bool no_amb = false;
+    bool trim_paths = false;
 
     if (argc < 3){
         cerr << "No GFA file given as input." << endl << endl;
         trim_help(argv);
         exit(1);
     }
+    int c;
+    optind = 2;
+    while (true){
+        static struct option long_options[] =
+        {
+            {"help", no_argument, 0, 'h'},
+            {"trim-paths", no_argument, 0, 'p'},
+            {"length", required_argument, 0, 'l'},
+            {"no-ambiguous", no_argument, 0, 'N'},
+            {"version", no_argument, 0, 'v'},
+            {0,0,0,0}
+        };
+    
+        int option_index = 0;
+        c = getopt_long(argc, argv, "hvnpl:", long_options, &option_index);
+        if (c == -1){
+            break;
+        }
 
+        switch (c){
+            case '?':
+            case 'h':
+                extract_help(argv);
+                exit(0);
+            case 'v':
+                print_version_help();
+                exit(0);
+            case 'l':
+                minlen = atoi(optarg);
+                break;
+            case 'N':
+                no_amb = true;
+                break;
+            case 'p':
+                trim_paths = true;
+                cerr << "Path trimming is not yet implemented" << endl;
+                exit(1);
+            default:
+                abort();
+        }
+    }
 
+    if (optind > argc){
+        cerr << "Error: no GFA file provided." << endl;
+        exit(1);
+    }
+    gfa_file = argv[optind];
+    
+    GFAKluge gg;
+    gg.parse_gfa_file(gfa_file);
+    gg.trim_seqs(minlen, no_amb);
+
+    cout << gg;
 
     return 0;
 }
@@ -1136,6 +1191,9 @@ int main(int argc, char** argv){
     }
     else if (strcmp(argv[1], "build") == 0){
         return build_main(argc, argv);
+    }
+    else if (strcmp(argv[1], "trim") == 0){
+        return trim_main(argc, argv);
     }
     else {
         cerr << "No command " << '"' << argv[1] << '"' << endl;
