+++ src/gfakluge.cpp
@@ -92,6 +92,53 @@ namespace gfak{
         }
     }
 
+    bool GFAKluge::trim_seqs(const int& min_len, const bool& no_amb){
+        bool graph_modified = false;
+
+        unordered_set<string> dropped_seqs;
+        map<string, sequence_elem, custom_key>::iterator n_to_s;
+        for (n_to_s = name_to_seq.begin(); n_to_s != name_to_seq.end(); n_to_s++){
+            auto& s = n_to_s->second;
+            if (s.length == UINT64_MAX){
+                cerr << "Length unset for sequence " << s.name << "; removing from graph." << endl;
+                dropped_seqs.insert(s.name);
+                name_to_seq.erase(n_to_s);
+                graph_modified = true;
+            }
+            else if (min_len > 0 && s.length < min_len){
+                dropped_seqs.insert(s.name);
+                name_to_seq.erase(n_to_s);
+                graph_modified = true;
+            }
+
+            if (no_amb && !pliib::canonical(s.sequence)){
+                dropped_seqs.insert(s.name);
+                //name_to_seq.erase(n_to_s);
+                graph_modified = true;
+            }
+        }
+
+        map<string, vector<edge_elem>>::iterator s_to_e;
+        for (s_to_e = seq_to_edges.begin(); s_to_e != seq_to_edges.end(); s_to_e++){
+            // Remove all edges that have dropped sequences as their source.
+            if (dropped_seqs.find(s_to_e->first) != dropped_seqs.end()){
+                //seq_to_edges.erase(s_to_e);
+                graph_modified = true;
+                continue;
+            }
+            // Remove any individual edges that have a dropped sequence as their sink.
+            vector<edge_elem>::iterator ee;
+            for (ee = s_to_e->second.begin(); ee != s_to_e->second.end(); ee++){
+                if (dropped_seqs.find(ee->sink_name) != dropped_seqs.end()){
+                   // s_to_e->second.erase(ee);
+                    graph_modified = true;
+                }
+            }
+        }
+
+        return graph_modified;
+    }
+
 
     void GFAKluge::groups_as_paths(){
         for (auto g : groups){
