### 2018-10-04 71a976a
Fix trim_main help and opt parsing
