### 2018-10-04 d4cb5ea
Change the way edges get trimmed to prevent segfaulting due to pointer invalidation
