### 2018-10-12 bddd320
Fix fasta output in convert
