+++ src/main.cpp
@@ -534,7 +534,7 @@ int convert_main(int argc, char** argv){
 
     if (make_fasta){
         for (auto s : gg.get_name_to_seq()){
-            cout << s.second.as_fasta_record();
+            cout << s.second.as_fasta_record() << endl; 
         }
         exit(0);
     }
