### 2018-11-06 #975 @edawson
These proofs look good. Is there more I need to do to bump this towards review?

We'd like to suggest @ctb , @luizirber , @afrubin , @betteridiot  as possible reviewers based on the list of people who have volunteered previously.  @sjackman and @richarddurbin might also have good ideas about possible reviewers (maybe @jts ?). Thanks!
