### 2018-11-06 #975 @sjackman
Giorgio Gonnella @ggonnella, author of https://github.com/ggonnella/gfapy and https://github.com/ggonnella/rgfa, is a possible reviewer.
