### 2018-11-14 #975 @mgymrek
@ctb @afrubin @betteridiot would either of you be able to review this submission?
