### 2018-11-14 #975 @betteridiot
As interesting as this is, I am a reviewer for another JoSS paper right now.
