### 2018-11-16 #1083 @whedon
[REVIEW]: GFAKluge: A C++ library and command line utilities for the Graphical Fragment Assembly formats
**Submitting author:** @edawson (<a href="http://orcid.org/0000-0001-5448-1653">Eric Dawson</a>)
**Repository:** <a href="https://github.com/edawson/gfakluge" target ="_blank">https://github.com/edawson/gfakluge</a>
**Version:** 1.1.2
