### 2018-11-28 #1083 @sjackman
> Installation: Does installation proceed as outlined in the documentation?

> You can build libgfakluge and the command line gfak utilities by typing `make` in the repo.

On macOS with Apple `clang`, `make` fails with `error: 'omp.h' file not found`, because `gfakluge` requires OpenMP, and Apple `clang` does not provide OpenMP. Please add OpenMP to the list of dependencies. Please include instructions in `README.md` for building from source on macOS.

```sh
brew install gcc@8
make CXX=g++-8
```
