### 2018-11-28 #1083 @sjackman
> GFAKluge [@GFAKluge] is a set of command line utilities

> ```bibtex
@online{GFAKluge,
  author = {Eric T. Dawson and Richard Durbin},
  title = {GFAKluge},
  year = 2018,
  url = {https://github.com/edawson/gfakluge}
}
```

I don't believe it's typical to cite the software repository from the paper.

@mgymrek Opinion?
