### 2018-11-28 #1083 @sjackman
> GFAKluge facilitates interprogram exchange by providing a high-level C++ API for developers

> To use GFAKluge in your program, you'll need to add a few lines to your code. First, add the necessary include line to your C++ code: `#include "gfakluge.hpp"`
Next, make sure that the library is on the proper system paths and compile line:

GFAKluge would be more easily used as a library if

1. it were a header-only library (so it does not require compilation and `-L… -l…`)
2. `make install` and the Brew package installed that header

To make `gfakluge` a header-only library, merge `gfakluge.hpp` and `gfakluge.cpp` into a single file and remove `gfakluge.cpp`. Make every method an inline method by either defining it within `class GFAKluge` or decorating the definition with `static inline` if it's defined outside the class.

Header-only libraries are much easier to use in practice than libraries that require linking.
