### 2018-11-28 #1083 @sjackman
Please add a `make install` target to `Makefile`.

```makefile
prefix=/usr/local

install:
	install gfak $(DESTDIR)$(prefix)/bin/
```
