### 2018-11-28 #1083 @sjackman
> and converting between legacy GFA formats.

Please be more specific. I suggest `and converting between GFA versions 1 and 2.`
