### 2018-11-28 #1083 @sjackman
> To our knowledge, GFAKluge is the only publically-available software package that can consume and produce both GFA1 and GFA2

GfaPy is able to consume and produce both GFA1 and GFA2.
https://doi.org/10.1093/bioinformatics/btx398

RGFA handles only GFA1 (I believe). I suggest citing it.
https://doi.org/10.7717/peerj.2681

`abyss-todot` included with ABySS 2 produces both GFA1 and GFA2 (with the command line options `--gfa1` and `--gfa2`), and consumes both with some format limitations (it does not handle all record types).
https://doi.org/10.1101/gr.214346.116

See https://github.com/GFA-spec/GFA-spec#implementations
