### 2018-11-28 #1083 @sjackman
> the gfak convert tool

I'd suggest typesetting this as…

the `gfak convert` tool
