### 2018-11-28 #1083 @sjackman
> We see the command line utilities as being useful to the development community in the short term.

The need for file conversion will likely never go away. Older tools that support only GFA1 may never be updated to support GFA2. Newer tools may support producing only GFA2 and not GFA1. To make these tools interoperate, a conversion tool is needed. That need is likely not temporary.
