### 2018-11-28 #1083 @sjackman
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Please list in the manuscript the currently available commands.

```
   convert: Convert between GFA 0.1 <-> 1.0 <-> 2.0
   diff:    Determine whether two GFA files have identical graphs
   extract: Convert the S lines of a GFA file to FASTA format.
   fillseq: Add sequences from a FASTA file to S lines.
   ids:     Coordinate the ID spaces of multiple GFA graphs.
   concat:  Merge GFA graphs (without ID collisions).
   sort:    Print a GFA file in HSLP / HSEFGUO order.
   stats:   Get assembly statistics (e.g. N50) for a GFA file.
   subset:  Extract the subgraph between two IDs in a graph.
   trim:    Remove elements from a GFA graph.
```

Please give three or more real world examples of their use.
