### 2018-11-28 #1083 @sjackman
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Is the API documented? If not, please document the API using Doxygen or a similar tool. A one-line summary of the purpose of each public method is sufficient. Please post the rendered HTML API documentation online using GitHub Pages or similar.
