### 2018-11-28 #1083 @sjackman
>  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

Please include a `make check` rule in the Makefile that runs the automated tests.
