### 2018-11-28 #1083 @sjackman
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I've ticked this off as complete, since I think it's self evident how to use GitHub to submit issues and pull requests. You could if you liked add `Reporting bugs` and `Contributing` to the README.md or CONTRIBUTING.md.
