### 2018-11-30 #1083 @mgymrek
Thanks for these helpful comments @sjackman.

To summarize for @edawson:

- Please address the comments regarding installation and documentation.
- I agree no need to cite the software repository in the paper.
- There is no need to include extensive documents of the different commands in the manuscript itself. However please be sure these are well documented e.g. in the readme or manual.
- For completeness it would be helpful to add a CONTRIBUTING.md file.
