### 2018-12-05 12fa211
Update the build instructions to include OpenMP dependency
