### 2018-12-10 cf38c18
Update build instructions in readme
