### 2018-12-10 5d60030
Fix build on OS X Mojave
