+++ src/gfakluge.cpp
@@ -6,49 +6,6 @@
 using namespace std;
 namespace gfak{
 
-    int determine_line_type(char* line){
-       if (line[0] == 'H'){
-            return HEADER_LINE;
-       }
-       else if (line[0] == 'S'){
-            return SEGMENT_LINE;
-       }
-       else if (line[0] == 'E' || line[0] == 'L' || line[0] == 'C'){
-            return EDGE_LINE;
-       }
-       else if (line[0] == 'U' || line[0] == 'O' || line[0] == 'P' || line[0] == 'W'){
-            return GROUP_LINE;
-       }
-       else if (line[0] == 'G'){
-            return GAP_LINE;
-       }
-       else if (line[0] == 'F'){
-            return FRAGMENT_LINE;
-       }
-       else{
-            return -1;
-       }
-    }
-
-    map<string, vector<edge_elem>> GFAKluge::get_seq_to_edges(){
-        return seq_to_edges;
-    }
-    map<string, vector<fragment_elem>> GFAKluge::get_seq_to_fragments(){
-        return seq_to_fragments;
-    }
-    map<string, vector<gap_elem>> GFAKluge::get_seq_to_gaps(){
-        return seq_to_gaps;
-    }
-    map<string, group_elem> GFAKluge::get_groups(){
-        return groups;
-    }
-
-    void GFAKluge::add_edge(const string& seqname, const edge_elem& e){
-        seq_to_edges[seqname].push_back(e);
-    }
-    void GFAKluge::add_edge(const sequence_elem& s, const edge_elem& e){
-        add_edge(s.name, e);
-    }
 
     void GFAKluge::add_fragment(const string& seqname, const fragment_elem& f){
         seq_to_fragments[seqname].push_back(f);
@@ -174,161 +131,6 @@ namespace gfak{
         gfa_2_ize();
     }
 
-    void GFAKluge::gfa_1_ize(){
-        if (!one_compat){
-
-        
-        /**
-         * Set base_edge_id to max_seq_id + 1
-        * Swap edges to links
-        * Ordered groups -> paths,
-        * warn about missing sets
-        * warn about missing fragments (we could output them as seqs)
-        * warn about missing gaps (we could remove them)
-        * 
-        */
-            string k = name_to_seq.rbegin()->first;
-            if (string_is_number(k)){
-                base_edge_id = std::stoul(k) + 1;
-            }
-            else{
-                base_edge_id = name_to_seq.size();
-            }
-            groups_as_paths();
-            for (auto s = name_to_seq.begin(); s != name_to_seq.end(); s++){
-                if (s->second.sequence != "*"){
-                    s->second.length = s->second.sequence.length();
-                }
-                for (auto e = seq_to_edges[s->first].begin(); e != seq_to_edges[s->first].end(); e++){
-                    int t = e->determine_type();
-                    if (e->id == "*"){
-                        e->id = std::to_string(++base_edge_id);
-                    }
-                    if (t == 1){
-                        e->ends.set(0,1);
-                        e->ends.set(1,1);
-                        e->ends.set(2,0);
-                        e->ends.set(3,0);            
-                        e->source_begin = s->second.length;
-                        e->source_end = s->second.length;
-                        e->sink_begin = 0;
-                        e->sink_end = 0;
-                        if (e->id == "*"){
-                            e->id = std::to_string(++base_edge_id);
-                        }
-                    }
-                    else if (t == 2){
-                        sequence_elem sink = name_to_seq[e->sink_name];
-                        e->ends.set(0,(e->source_begin == s->second.length));
-                        e->ends.set(1,(e->source_end == s->second.length));
-                        e->ends.set(2,(e->sink_begin == sink.length));
[…]
