+++ src/gfakluge.hpp
@@ -22,33 +22,30 @@ namespace gfak{
 
     enum gfa_line_types {HEADER_LINE,SEGMENT_LINE,FRAGMENT_LINE,EDGE_LINE,GAP_LINE,GROUP_LINE,PATH_LINE};
 
-    inline int detect_line_type(const string& line){
-        char c = line[0];
-        if (c == 'S'){
+    inline int determine_line_type(char* line){
+        if (line[0] == 'H'){
+            return HEADER_LINE;
+        }
+        else if (line[0] == 'S'){
             return SEGMENT_LINE;
         }
-        else if (c == 'E' || c == 'L' || c == 'C'){
+        else if (line[0] == 'E' || line[0] == 'L' || line[0] == 'C'){
             return EDGE_LINE;
         }
-        else if (c == 'P'){
-            return PATH_LINE;
-        }
-        else if (c == 'F'){
-            return FRAGMENT_LINE;
-        }
-        else if (c == 'G'){
-            return GAP_LINE;
-        }
-        else if (c == 'O' || c == 'U'){
+        else if (line[0] == 'U' || line[0] == 'O' || line[0] == 'P' || line[0] == 'W'){
             return GROUP_LINE;
         }
+        else if (line[0] == 'G'){
+            return GAP_LINE;
+        }
+        else if (line[0] == 'F'){
+            return FRAGMENT_LINE;
+        }
         else{
-            cerr << "Unknown GFA line type: " << line << endl;
             return -1;
         }
-        
     };
-    
+
 
     // Provides the proper sorting behavior,
     // where number-based keys get sorted in numerical order
@@ -81,7 +78,7 @@ namespace gfak{
         uint64_t num_fragments = 0;
         uint64_t num_gaps = 0;
         uint64_t num_paths = 0;
-        
+
         double N50 = 0.0;
         double N90 = 0.0;
     };
@@ -115,9 +112,9 @@ namespace gfak{
 
     /** 
      * Encodes an annotation line.
-    *  These begin with an 'x', and are not part of the official
-    *  community GFA spec, but have appeared in early GFA files.
-    */
+     *  These begin with an 'x', and are not part of the official
+     *  community GFA spec, but have appeared in early GFA files.
+     */
     struct annotation_elem{
         std::string key;
         std::string info;
@@ -198,9 +195,9 @@ namespace gfak{
                 ++rank;
                 st << 'W' << '\t' << this->segment_names[i] << '\t' << 
                     this->name << '\t' << rank << '\t' <<
-                     (this->orientations[i] ? "+" : "-");
+                    (this->orientations[i] ? "+" : "-");
                 if (this->overlaps.size() == this->segment_names.size()){
-                   st << '\t' << overlaps[i];
+                    st << '\t' << overlaps[i];
                 }
                 st << endl;
                 os << st.str();
@@ -321,35 +318,35 @@ namespace gfak{
     struct edge_elem{
 
         edge_elem(){
-            
+
         }
         edge_elem(const link_elem& l){
-           source_name = l.source_name;
-           sink_name = l.sink_name;
-           type = 1;
-           source_orientation_forward = l.source_orientation_forward;
-           sink_orientation_forward = l.source_orientation_forward;
-           alignment = l.cigar;
-           tags = l.opt_fields;
-           ends.set(0,1);
-           ends.set(1,1);
-           ends.set(2,0);
-           ends.set(3,0);
-           id = "*";
+            source_name = l.source_name;
+            sink_name = l.sink_name;
+            type = 1;
+            source_orientation_forward = l.source_orientation_forward;
+            sink_orientation_forward = l.source_orientation_forward;
+            alignment = l.cigar;
+            tags = l.opt_fields;
+            ends.set(0,1);
[…]
