### 2018-12-11 fe0ee64
Mid header-only-ification
