+++ src/gfakluge.hpp
@@ -542,6 +542,55 @@ namespace gfak{
         class GFAKluge{
             friend std::ostream& operator<<(std::ostream& os, GFAKluge& g);
 
+
+
+
+            private:
+            bool use_walks = false;
+            // Store whether we've already gone walks->paths and paths->walks
+            bool normalized_paths = false;
+            bool normalized_walks = false;
+
+            bool one_compat = false;
+            bool two_compat = false;
+
+            uint64_t next_set_or_path_id = 0;
+            uint64_t base_seq_id = 0;
+            uint64_t base_edge_id = 0;
+            uint64_t base_gap_id = 0;
+            uint64_t base_frag_id = 0;
+            uint64_t base_group_id = 0;
+
+
+            double version = 0.0;
+            std::map<std::string, header_elem> header;
+            std::map<std::string, std::vector<contained_elem> > seq_to_contained;
+            std::map<std::string, std::vector<link_elem> > seq_to_link;
+            //Since we can't compare sequence elements for hashing,
+            // we cheat and use their names (which are only sort of guaranteed to be
+            // unique.
+            std::map<std::string, sequence_elem, custom_key> name_to_seq;
+            bool string_is_number(std::string s);
+            std::string join(const std::vector<std::string>& splits, const std::string& glue);
+            std::map<std::string, std::vector<alignment_elem> > seq_to_alignment;
+            std::string header_string(std::map<std::string, header_elem>& opts);
+            std::string opt_string(std::vector<opt_elem> opts);
+
+            std::map<std::string, path_elem> name_to_path;
+
+            /** GFA 2.0 containers **/
+            std::map<std::string, std::vector<fragment_elem> > seq_to_fragments;
+            std::map<std::string, std::vector<edge_elem> > seq_to_edges;
+            std::map<std::string, std::vector<gap_elem> > seq_to_gaps;
+            std::map<std::string, group_elem> groups;
+
+            double n50 = -1;
+            double l50 = -1;
+            double n90 = -1;
+            double l90 = -1;
+            double simple_connectivity = -1;
+            double weighted_connectivity = -1;
+
             public:
             GFAKluge(){
                 map<std::string, header_elem> header;
@@ -565,6 +614,419 @@ namespace gfak{
             ~GFAKluge(){
 
             };
+
+            // GFAKluge does not, but maybe should, enforce graph structure,
+            // i.e.:
+            // 1. Throw errors on dangling edges
+            // 2. Guarantee contained elements fall in valid sequences
+
+
+            /**
+             * GFA2.0 handlers
+             */
+            void add_sequence(sequence_elem s){
+                name_to_seq[s.name] = s;
+            }
+            void add_edge(const std::string& seqname, const edge_elem& e){
+                seq_to_edges[seqname].push_back(e);
+            }
+            void add_edge(const sequence_elem& s, const edge_elem& e){
+                add_edge(s.name, e);
+            }
+
+            void add_fragment(const std::string& seqname, const fragment_elem& f){
+        seq_to_fragments[seqname].push_back(f);
+    }
+            
+            void add_fragment(const sequence_elem& s, const fragment_elem& f){
+        add_fragment(s.name, f);
+    }
+
+            void add_gap(const gap_elem& g){
+        seq_to_gaps[g.source_name].push_back(g);
+    }
+
+            void add_group(const group_elem& g){
+        this->groups[g.id] = g;
+    }
+
+            // Convert group_elems (i.e. U or O lines) to path_elems (P lines)
+            void groups_as_paths(){
+        for (auto g : groups){
+            if (g.second.ordered){
+                path_elem p;
+                p.name = g.first;
+                p.segment_names = g.second.items;
+                if (this->name_to_seq.size() > 0){
[…]
