+++ src/gfakluge.cpp
@@ -7,21 +7,7 @@ using namespace std;
 namespace gfak{
 
 
-    void GFAKluge::add_fragment(const string& seqname, const fragment_elem& f){
-        seq_to_fragments[seqname].push_back(f);
-    }
 
-    void GFAKluge::add_fragment(const sequence_elem& s, const fragment_elem& f){
-        add_fragment(s.name, f);
-    }
-
-    void GFAKluge::add_gap(const gap_elem& g){
-        seq_to_gaps[g.source_name].push_back(g);
-    }
-
-    void GFAKluge::add_group(const group_elem& g){
-        this->groups[g.id] = g;
-    }
 
     void GFAKluge::fill_sequences(const char* fasta_file){
 
@@ -99,59 +85,10 @@ namespace gfak{
     }
 
 
-    void GFAKluge::groups_as_paths(){
-        for (auto g : groups){
-            if (g.second.ordered){
-                path_elem p;
-                p.name = g.first;
-                p.segment_names = g.second.items;
-                if (this->name_to_seq.size() > 0){
-                    p.overlaps.resize(p.segment_names.size());
-                    for (size_t i = 0; i < p.segment_names.size(); ++i){
-                        int len = 0;
-                        if (name_to_seq.find(p.segment_names[i]) != name_to_seq.end()){
-                            std::string s(name_to_seq.at(p.segment_names[i]).sequence);
-                            len = s.length();
-                        }
-                        p.overlaps[i].assign(std::to_string(len) + "M");
-                    }
-                }
-                p.orientations = g.second.orientations;
-                p.opt_fields = g.second.tags;
-                add_path(p.name, p);
-            }
-            else{
-                cerr << "Group " << g.first << " is unordered; skipping adding it to the paths." << endl;
-            }
-        }
-    }
 
     void GFAKluge::compatibilize(){
         gfa_1_ize();
         gfa_2_ize();
-    }
-
-    double GFAKluge::get_version(){
-        return this->version;
-    }
-
-    void GFAKluge::set_version(double v){
-        header_elem verz;
-        verz.key = "VN";
-        verz.type="Z";
-        this->version = v;
-        verz.val = std::to_string((double) this->version).substr(0,3);
-        this->header[verz.key] = verz;
-        //gfa_1_ize();
-        //gfa_2_ize();
-    }
-    void GFAKluge::set_version(){
-        header_elem verz;
-        verz.key = "VN";
-        verz.type="Z";
-        verz.val = std::to_string((double) this->version).substr(0,3);
-
-        this->header[verz.key] = verz;
     }
 
     bool GFAKluge::string_is_number(string s){
@@ -166,60 +103,6 @@ namespace gfak{
         return !(s.empty());
     }
 
-
-    void GFAKluge::add_path(string path_name, path_elem p){
-        name_to_path[path_name] = p;
-    }
-
-    void GFAKluge::add_walk(std::string pathname, const int& rank, const string& segname, const bool& ori, const string& overlap, vector<opt_elem> opts){
-        if (name_to_path.find(pathname) == name_to_path.end()){
-            path_elem p;
-            p.name = pathname;
-            add_path(p.name, p);
-        }
-        name_to_path.at(pathname).add_ranked_segment( rank, segname, ori, overlap, opts);
-
-    }
-
-    void GFAKluge::add_link(const sequence_elem& seq, const link_elem& link){
-        edge_elem e(link);
-        seq_to_edges[seq.name].push_back(e);
-    }
-
-    void GFAKluge::add_contained(sequence_elem seq, contained_elem con){
[…]
