+++ src/gfakluge.hpp
@@ -11,6 +11,7 @@
 #include <cstdlib>
 #include <iostream>
 #include <bitset>
+#include <unordered_set>
 #include <sys/stat.h>
 
 #include "tinyfa.hpp"
@@ -540,7 +541,14 @@ namespace gfak{
         };
 
         class GFAKluge{
-            friend std::ostream& operator<<(std::ostream& os, GFAKluge& g);
+            friend std::ostream& operator<<(std::ostream& os, GFAKluge& g){
+        g.gfa_1_ize();
+        g.gfa_2_ize();
+
+        //os << g.to_string();
+        g.output_to_stream(os);
+        return os;
+    }
 
 
 
@@ -570,11 +578,57 @@ namespace gfak{
             // we cheat and use their names (which are only sort of guaranteed to be
             // unique.
             std::map<std::string, sequence_elem, custom_key> name_to_seq;
-            bool string_is_number(std::string s);
-            std::string join(const std::vector<std::string>& splits, const std::string& glue);
+            
+            bool string_is_number(std::string s){
+        bool ret = true;
+        std::string::iterator it;
+        for (it = s.begin(); it != s.end(); it++){
+            if (!isdigit(*it)){
+                ret = false;
+                return ret;
+            }
+        }
+        return !(s.empty());
+    }
+            
+            std::string join(const std::vector<std::string>& splits, const std::string& glue){
+        string ret = "";
+        for (size_t i = 0; i < splits.size(); i++){
+            if (i != 0){
+                ret += glue;
+            }
+            ret += splits[i];
+        }
+
+        return ret;
+    }
             std::map<std::string, std::vector<alignment_elem> > seq_to_alignment;
-            std::string header_string(std::map<std::string, header_elem>& opts);
-            std::string opt_string(std::vector<opt_elem> opts);
+            std::string header_string(std::map<std::string, header_elem>& headers){
+        string ret = "H";
+        map<string, header_elem>::iterator it;
+        for (it = headers.begin(); it != headers.end(); it++){
+            ret += "\t";
+            header_elem h = it->second;
+            string t[] = {h.key, h.type, h.val};
+            vector<string> temp = vector<string> (t, t + sizeof(t) / sizeof(string));
+            ret += join(temp, ":");
+        }
+        return ret;
+
+    }
+            std::string opt_string(std::vector<opt_elem> opts){
+        string ret = "";
+        for (size_t i = 0; i < opts.size(); i++){
+            opt_elem o = opts[i];
+            if (i != 0){
+                ret += "\t";
+            }
+            string t [] = {o.key, o.type, o.val};
+            vector<string> temp = vector<string> (t, t + sizeof(t) / sizeof(string));
+            ret += join(temp, ":");
+        }
+        return ret;
+    }
 
             std::map<std::string, path_elem> name_to_path;
 
@@ -766,6 +820,55 @@ namespace gfak{
         this->use_walks = ws;
     }
 
+                /** Methods for folks that want streaming output.
+             *  Writes a single element to an ostream, using either
+             *  that element's to_string_1() or to_string_2() function
+             *  depending on the version set in the GFAKluge object.
+             */
+            void write_element(std::ostream& os, const sequence_elem& s) {
+        if (this->version >= 2.0){
+            os << s.to_string_2();
+        }
+        else{
+            os << s.to_string_1();
+        }
+    }
+            void write_element(std::ostream& os, edge_elem e){
+        if (this->version >= 2.0){
+            os << e.to_string_2();
+        }
+        else{
+            os << e.to_string_1();
+        }
+    }
+            void write_element(std::ostream& os, const fragment_elem& f){
+        os << f.to_string();
+    }
[…]
