+++ src/gfakluge.cpp
@@ -1,851 +1,0 @@
-#include "gfakluge.hpp"
-
-#include <unordered_set>
-#include <fstream>
-
-using namespace std;
-namespace gfak{
-
-
-
-
-    void GFAKluge::fill_sequences(const char* fasta_file){
-
-        TFA::tiny_faidx_t tf;
-        if (TFA::checkFAIndexFileExists(fasta_file)){
-            TFA::parseFAIndex(fasta_file, tf);
-        }
-        else{
-            cerr << "Creating index for " << fasta_file << "." << endl;
-            TFA::createFAIndex(fasta_file, tf);
-            cerr << "Created index." << endl;
-            TFA::writeFAIndex(fasta_file, tf);
-            cerr << "Wrote index to file." << endl;
-        }
-        
-        for (std::map<string, sequence_elem, custom_key>::iterator it = name_to_seq.begin(); it != name_to_seq.end(); it++){
-            if (tf.hasSeqID(it->second.name.c_str())){
-                char* s;
-                TFA::getSequence(tf, it->second.name.c_str(), s);
-                it->second.sequence.assign(s);
-                // The length field should already be filled, but it
-                // might be good to check.
-                delete [] s;
-            }
-        }
-    }
-
-    bool GFAKluge::trim_seqs(const int& min_len, const bool& no_amb){
-        bool graph_modified = false;
-
-        unordered_set<string> dropped_seqs;
-        map<string, sequence_elem, custom_key>::iterator n_to_s;
-        for (n_to_s = name_to_seq.begin(); n_to_s != name_to_seq.end(); n_to_s++){
-            auto& s = n_to_s->second;
-            if (s.length == UINT64_MAX){
-                cerr << "Length unset for sequence " << s.name << "; removing from graph." << endl;
-                dropped_seqs.insert(s.name);
-                name_to_seq.erase(n_to_s);
-                graph_modified = true;
-            }
-            else if (min_len > 0 && s.length < min_len){
-                dropped_seqs.insert(s.name);
-                name_to_seq.erase(n_to_s);
-                graph_modified = true;
-            }
-
-            if (no_amb && !pliib::canonical(s.sequence)){
-                dropped_seqs.insert(s.name);
-                name_to_seq.erase(n_to_s);
-                graph_modified = true;
-            }
-        }
-
-        map<string, vector<edge_elem>>::iterator s_to_e;
-        for (s_to_e = seq_to_edges.begin(); s_to_e != seq_to_edges.end(); s_to_e++){
-            // Remove all edges that have dropped sequences as their source.
-            if (dropped_seqs.find(s_to_e->first) != dropped_seqs.end()){
-                seq_to_edges.erase(s_to_e);
-                graph_modified = true;
-                continue;
-            }
-            // Remove any individual edges that have a dropped sequence as their sink.
-            vector<edge_elem> pass_elems;
-            for (auto ee : s_to_e->second){
-                if (dropped_seqs.find(ee.source_name) == dropped_seqs.end()){
-                    graph_modified = true;
-                    pass_elems.push_back(ee);
-                }
-            }
-            seq_to_edges[s_to_e->first] = pass_elems;
-        }
-
-
-        return graph_modified;
-    }
-
-
-
-    void GFAKluge::compatibilize(){
-        gfa_1_ize();
-        gfa_2_ize();
-    }
-
-    bool GFAKluge::string_is_number(string s){
-        bool ret = true;
-        std::string::iterator it;
-        for (it = s.begin(); it != s.end(); it++){
-            if (!isdigit(*it)){
-                ret = false;
-                return ret;
-            }
-        }
-        return !(s.empty());
-    }
-
-
-    map<string, vector<link_elem> > GFAKluge::get_seq_to_link(){
-        for (auto s : name_to_seq){
[…]

[…]
