### 2018-12-13 4ffcfcd
Change typesetting. Add make install target.
