### 2018-12-14 8ca1d2c
Add streaming callback functions
