+++ src/gfakluge.hpp
@@ -5,6 +5,7 @@
 #include <sstream>
 #include <istream>
 #include <algorithm>
+#include <functional>
 #include <map>
 #include <vector>
 #include <sstream>
@@ -21,20 +22,32 @@ namespace gfak{
 
     // ( <header> | <segment> | <fragment> | <edge> | <gap> | <group> )
 
-    enum gfa_line_types {HEADER_LINE,SEGMENT_LINE,FRAGMENT_LINE,EDGE_LINE,GAP_LINE,GROUP_LINE,PATH_LINE};
+    enum gfa_line_types {HEADER_LINE,SEGMENT_LINE,FRAGMENT_LINE,EDGE_LINE,GAP_LINE,GROUP_LINE,PATH_LINE,LINK_LINE,CONTAINED_LINE,WALK_LINE};
 
-    static inline int determine_line_type(char* line){
+    static inline int determine_line_type(const char* line){
         if (line[0] == 'H'){
             return HEADER_LINE;
         }
         else if (line[0] == 'S'){
             return SEGMENT_LINE;
         }
-        else if (line[0] == 'E' || line[0] == 'L' || line[0] == 'C'){
+        else if (line[0] == 'E'){
             return EDGE_LINE;
         }
-        else if (line[0] == 'U' || line[0] == 'O' || line[0] == 'P' || line[0] == 'W'){
+        else if (line[0] == 'L'){
+            return LINK_LINE;
+        }
+        else if (line[0] == 'C'){
+            return CONTAINED_LINE;
+        }
+        else if (line[0] == 'U' || line[0] == 'O'){
             return GROUP_LINE;
+        }
+        else if (line[0] == 'P'){
+            return PATH_LINE;
+        }
+        else if (line[0] == 'W'){
+            return WALK_LINE;
         }
         else if (line[0] == 'G'){
             return GAP_LINE;
@@ -542,16 +555,13 @@ namespace gfak{
 
         class GFAKluge{
             friend std::ostream& operator<<(std::ostream& os, GFAKluge& g){
-        g.gfa_1_ize();
-        g.gfa_2_ize();
+                g.gfa_1_ize();
+                g.gfa_2_ize();
 
-        //os << g.to_string();
-        g.output_to_stream(os);
-        return os;
-    }
-
-
-
+                //os << g.to_string();
+                g.output_to_stream(os);
+                return os;
+            }
 
             private:
             bool use_walks = false;
@@ -668,6 +678,255 @@ namespace gfak{
             ~GFAKluge(){
 
             };
+
+            double detect_version_from_file(const char* filename){
+                ifstream gfi;
+                gfi.open(filename, std::ifstream::in);
+                if (!gfi.good()){
+                    cerr << "Couldn't open GFA file " << filename << "." << endl;
+                    exit(1);
+                }
+                string line;
+                while (getline(gfi, line)){
+                    if (determine_line_type(line.c_str()) == HEADER_LINE){
+                        vector<string> tokens = pliib::split(line, '\t');
+                        header_elem h;
+                        vector<string> line_tokens = pliib::split(tokens[1], ':');
+                        //TODO this is not well implemented
+                        // GFA places no guarantees on header format
+                        h.key = line_tokens[0];
+                        h.type = line_tokens[1];
+                        h.val = line_tokens[2];
+                        if (h.key.compare("VN") == 0){
+                            set_version(stod(h.val));
+                            break;
+                        }
+                        header[h.key] = h;
+                    } 
+                }
+            };
+
+            void for_each_sequence_line_in_file(const char* filename, std::function<void(gfak::sequence_elem)> func){
+                ifstream gfi;
+                gfi.open(filename, std::ifstream::in);
+                if (!gfi.good()){
+                    cerr << "Couldn't open GFA file " << filename << "." << endl;
[…]
