### 2018-12-28 bc6dc37
hack in edge and link streaming parsing
