### 2018-12-30 5a3094a
Update the paper with changes suggested in the review.
