### 2019-01-14 e197003
Add contributing section to README
