### 2019-01-14 7f53b58
Add command line examples to paper
