### 2019-01-14 4cc0faf
Add make check to makefile. Add bib entries for RGFA, GfaPy, Abyss
