+++ paper/paper.bib
@@ -52,4 +52,32 @@
     url = {https://github.com/GFA-spec/GFA-spec},
     year = {2015}
 }
+@article{GfaPy,
+    author = {Giorgio Gonnella and Stefan Kurtz},
+    title = {GfaPy: a flexible and extensible software library for handling sequence graphs in Python},
+    journal = {Bioinformatics},
+    volume = {33},
+    pages = {3094-3095},
+    doi = {doi:10.1093/bioinformatics/btx398},
+    year = {2017}
+
+}
+@article{RGFA,
+    author = {Giorgio Gonnella, Stefan Kurtz},
+    title = {RGFA: powerful and convenient handling of assembly graphs},
+    journal = {PeerJ},
+    volume = {4},
+    doi = {doi:10.7717/peerj.2681},
+    year = {2016}
+}
+@article{ABySS2.0,
+    author = {Shaun D. Jackman, Benjamin P. Vandervalk, Hamid Mohamadi, Justin Chu, Sarah Yeo, S. Austin Hammond, Golnaz Jahesh, Hamza Khan, Lauren Coombe, Rene L. Warren, Inanc Birol},
+    title = {ABySS 2.0: resource-efficient assembly of large genomes using a Bloom filter},
+    journal = {Genome Research},
+    volume = {27},
+    pages = {768-777},
+    doi = {doi:10.1101/gr.214346.116},
+    year = {2017}
+}
+
 
