### 2019-01-14 #1083 @edawson
@mgymrek I believe I've just addressed the last of these. I've included my checklist, which I pulled out from @sjackman 's comments.

- [X] Fix install instructions
- [X] Remove citation to repo in summary
- [X] Mention the homepage and license in the summary
- [X] Make GFAKluge a header-only library
- [X] Add a make install target to makefile
- [X] Change the wording of "legacy GFA formats"
- [X] Reference GFApy and RGFA and ABYSS, all of which are other programs which can convert GFA 1 -> 2
- [X] Fix typesetting of convert
- [X] Fix typesetting of vg msga
- [X] Change wording to reflect that GFA2 and GFA1 interconversion needs are not temporary
- [X] List the currently available commands in the manuscript.
- [X] Give command line examples of three or more commands
- [X] Mention the API documentation in the manuscript
- [X] Add a make test target in the makefile that runs automated tests
- [X] Add a contributing section to the Readme
