### 2019-01-16 fa0d0b6
Makefile (install): Run mkdir
