### 2019-01-16 #1083 @sjackman
For the authors of the reference "Gene Myers, Jason Chin, & Durbin, R. (2015)", I suggest adding Shaun Jackman (myself), Heng Li, and Giorgio Gonnella based on this text:

> Jason Chin, Richard Durbin, and myself (Gene Myers) found ourselves together at a workshop meeting in Dagstuhl Germany and hammered out an initial proposal for an assembly format. We started with GFA 1 and proceeded to build a more comprehensive design around it. After extensive revision and discussion on Github with the GFA group including Shaun Jackman, Heng Li, and Giorgio Gonnella, we arrived at GFA 2.0.

https://github.com/gfa-spec/gfa-spec/#gfa-20-graphical-fragment-assembly-gfa2-format-specification-20
