### 2019-01-16 #1083 @sjackman
> convert: Convert between GFA 0.1 <-> 1.0 <-> 2.0

What is GFA 0.1, and how does it differ from GFA 1.0?
