### 2019-01-21 #1083 @edawson
@sjackman I've addressed your remaining comments. We consider the original GFA proposal by Heng Li to be "GFA 0.1," and it's supported as a legacy format that predates any real specification. Thanks for your review!
