### 2019-01-21 #1083 @sjackman
> We consider the original GFA proposal by Heng Li to be "GFA 0.1," and it's supported as a legacy format that predates any real specification.

Are there any tangible technical differences between Heng's proposal GFA 0.1 and GFA 1.0?
