### 2019-01-21 #1083 @edawson
Off the top of my head, the only things that come to mind are the `x` and `a` lines sometimes seen in early minimap2 runs and the `W` walk line used in place of the `P` path line syntax.
