### 2019-01-21 #1083 @sjackman
> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

```console
$ make check
prove test/gfa_test.t
test/gfa_test.t .. test/gfa_test.t: line 5: ./bash-tap/bash-tap-bootstrap: No such file or directory
test/gfa_test.t: line 7: plan: command not found
test/gfa_test.t: line 10: is: command not found
test/gfa_test.t: line 14: is: command not found
test/gfa_test.t: line 17: is: command not found
test/gfa_test.t: line 20: is: command not found
test/gfa_test.t: line 23: is: command not found
test/gfa_test.t: line 26: is: command not found
Processed 1 graphs...
Done.
test/gfa_test.t: line 29: is: command not found
Merging 2 graphs...
test/gfa_test.t: line 32: is: command not found
test/gfa_test.t .. Dubious, test returned 127 (wstat 32512, 0x7f00)
No subtests run

Test Summary Report
-------------------
test/gfa_test.t (Wstat: 32512 Tests: 0 Failed: 0)
  Non-zero exit status: 127
  Parse errors: No plan found in TAP output
Files=1, Tests=0,  1 wallclock secs ( 0.04 usr  0.01 sys +  0.06 cusr  0.06 csys =  0.17 CPU)
Result: FAIL
make: *** [Makefile:40: check] Error 1
```

Should the file `bash-tap/bash-tap-bootstrap` be included in the source code, or are further instructions needed for the automated tests?
