### 2019-01-21 #1083 @sjackman
You may want to consider using Autotools (`autoconf` and `automake`) if you're interested. Not a requirement though.
