### 2019-01-21 #1083 @sjackman
`make check` fails on macOS with `md5sum: command not found`.
```console
$ make check
Last 15 lines from /Users/sjackman/Library/Logs/Homebrew/gfakluge/02.make:
test/gfa_test.t: line 32: md5sum: command not found
Merging 2 graphs...
test/gfa_test.t: line 32: is: command not found
test/gfa_test.t ..
Dubious, test returned 127 (wstat 32512, 0x7f00)
```
