+++ paper/paper.bib
@@ -47,7 +47,7 @@
     url = {http://lh3.github.io/2014/07/19/a-proposal-of-the-grapical-fragment-assembly-format}
 }
 @online{GFA-SPEC,
-    author = {Gene Myers, Jason Chin, and Richard Durbin},
+    author = {Gene Myers, Shaun Jackman, Giogio Gonnella, Jason Chin, and Richard Durbin},
     title = {Graphical Fragment Assembly (GFA) Format Specification},
     url = {https://github.com/GFA-spec/GFA-spec},
     year = {2015}
