### 2019-01-21 #1083 @edawson
- [X] make examples.md a link to that file
- [X] set the GitHub website link to the appropriate documentation (now links to README / GitHub Pages)
- [X] make interface.md a link to that file
- [X] bash-tap/bash-tap-bootstrap be included in the source code (made bash-tap) a git submodule, meaning it's now downloaded with `git clone --recursive`
- [X] You may want to consider using Autotools (I'm going to run through some tutorials and see if I can incorporate this - thanks for the suggestion!)
- [X] `make` install installs `pliib.hpp` and `tinyfa.hpp`

> I've updated brewsci/bio/gfakluge to 1.1.0.

Awesome! I was trying to do this before but struggling. I'd definitely appreciate the help learning how to bump a version and cut a fresh release on homebrew.
