### 2019-01-21 #1083 @sjackman
> Awesome! I was trying to do this before but struggling. I'd definitely appreciate the help learning how to bump a version and cut a fresh release on homebrew.

Open an issue or PR next time over at Brewsci/bio, and if you run into any trouble, I'll be happy to walk you through it.
