### 2019-01-21 #1083 @sjackman
> Thanks @sjackman, if all the items have been completed could you just update the checklist above?

@mgymrek All checked off, Melissa! ✅
