### 2019-01-22 b1dbc29
Fix Make install instructions
