+++ Makefile
@@ -34,6 +34,8 @@ $(BIN_DIR):
 
 install: gfak
 	cp src/gfakluge.hpp $(DESTDIR)$(PREFIX)/include/
+	cp src/tinyFA/tinyfa.hpp $(DESTDIR)$(PREFIX)/include/
+	cp src/tinyFA/pliib.hpp $(DESTDIR)$(PREFIX)/include/
 	cp gfak $(DESTDIR)$(PREFIX)/bin
 
 check : gfak
