### 2019-01-22 63b197b
Update Zenodo DOI badge with latest release
