+++ README.md
@@ -1,7 +1,7 @@
 gfakluge
 --------------------
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1434136.svg)](https://doi.org/10.5281/zenodo.1434136)
+[![DOI](https://zenodo.org/badge/47132080.svg)](https://zenodo.org/badge/latestdoi/47132080)
 
 ## What is it?  
 GFAKluge is a C++ parser/writer and
