### 2019-01-22 e254dac
Fix a typo in the bibliography
