### 2019-01-22 40c1070
Attempt to fix bib format wanting 'and' over commas
