+++ paper/paper.bib
@@ -5,7 +5,7 @@
   url = {https://github.com/edawson/gfakluge}
 }
 @article{TwoPaCo,
-    author = {Ilia Minkin, Son Pham, and Paul Medvedev},
+    author = {Ilia Minkin and Son Pham and Paul Medvedev},
     title = {TwoPaCo: An efficient algorithm to build the compacted de Bruijn graph from many complete genomes},
     journal = {Bioinformatics},
     volume = {33},
@@ -15,7 +15,7 @@
 }
 @article{Falcon,
     title = {Phased diploid genome assembly with single-molecule real-time sequencing},
-    author = {Chen-Shan Chin, Paul Peluso, Fritz J Sedlazeck, Maria Nattestad, Gregory T Concepcion, Alicia Clum, Christopher Dunn, Ronan O'Malley, Rosa Figueroa-Balderas, Abraham Morales-Cruz, Grant R Cramer, Massimo Delledonne, Chongyuan Luo, Joseph R Ecker, Dario Cantu, David R Rank and Michael C Schatz},
+    author = {Chen-Shan Chin and Paul Peluso and Fritz J Sedlazeck and Maria Nattestad and Gregory T Concepcion and Alicia Clum and Christopher Dunn and Ronan O'Malley and Rosa Figueroa-Balderas and Abraham Morales-Cruz and Grant R Cramer and Massimo Delledonne and Chongyuan Luo and Joseph R Ecker and Dario Cantu and David R Rank and Michael C Schatz},
     journal = {Nature Methods},
     year = {2016},
     volume = {13},
