### 2019-01-22 a74a7cd
Attempt to fix bib issue
