+++ paper/paper.bib
@@ -1,11 +1,11 @@
 @online{GFAKluge,
-  author = {Eric T. Dawson and Richard Durbin},
+  author = {Dawson, Eric T. and Durbin, Richard},
   title = {GFAKluge},
   year = 2018,
   url = {https://github.com/edawson/gfakluge}
 }
 @article{TwoPaCo,
-    author = {Ilia Minkin and Son Pham and Paul Medvedev},
+    author = {Minkin, Ilia and Pham, Son and Medvedev, Paul},
     title = {TwoPaCo: An efficient algorithm to build the compacted de Bruijn graph from many complete genomes},
     journal = {Bioinformatics},
     volume = {33},
@@ -16,6 +16,7 @@
 @article{Falcon,
     title = {Phased diploid genome assembly with single-molecule real-time sequencing},
     author = {Chen-Shan Chin and Paul Peluso and Fritz J Sedlazeck and Maria Nattestad and Gregory T Concepcion and Alicia Clum and Christopher Dunn and Ronan O'Malley and Rosa Figueroa-Balderas and Abraham Morales-Cruz and Grant R Cramer and Massimo Delledonne and Chongyuan Luo and Joseph R Ecker and Dario Cantu and David R Rank and Michael C Schatz},
+    author = {Chin, Chen-Shan and Peluso, Paul and Sedlazeck, Fritz J and Nattestad, Maria and Concepcion, Gregory T and Clum, Alicia and Dunn, Christopher and O'Malley, Ronan and Figueroa-Balderas, Rosa and Morales-Cruz, Abraham and Cramer, Grant R and Delledonne, Massimo and Luo, Chongyuan and Ecker, Joseph R and Cantu, Dario and Rank, David R and Schatz, Michael C},
     journal = {Nature Methods},
     year = {2016},
     volume = {13},
