### 2019-01-22 36b34bb
Tweak bibliography
