+++ paper/paper.bib
@@ -15,7 +15,6 @@
 }
 @article{Falcon,
     title = {Phased diploid genome assembly with single-molecule real-time sequencing},
-    author = {Chen-Shan Chin and Paul Peluso and Fritz J Sedlazeck and Maria Nattestad and Gregory T Concepcion and Alicia Clum and Christopher Dunn and Ronan O'Malley and Rosa Figueroa-Balderas and Abraham Morales-Cruz and Grant R Cramer and Massimo Delledonne and Chongyuan Luo and Joseph R Ecker and Dario Cantu and David R Rank and Michael C Schatz},
     author = {Chin, Chen-Shan and Peluso, Paul and Sedlazeck, Fritz J and Nattestad, Maria and Concepcion, Gregory T and Clum, Alicia and Dunn, Christopher and O'Malley, Ronan and Figueroa-Balderas, Rosa and Morales-Cruz, Abraham and Cramer, Grant R and Delledonne, Massimo and Luo, Chongyuan and Ecker, Joseph R and Cantu, Dario and Rank, David R and Schatz, Michael C},
     journal = {Nature Methods},
     year = {2016},
@@ -24,7 +23,7 @@
     doi = {doi:10.1038/nmeth.4035}
 }
 @article{vg,
-    author = {Erik Garrison, Jouni Siren, Adam M Novak, Glenn Hickey, Jordan M Eizinga, Eric T Dawson, William Jones, Michael F Lin, Benedict Paten, Richard Durbin},
+    author = {Garrison, Erik and  Sir\'{e}n, Jouni and Novak, Adam M and Hickey, Glenn and Eizinga, Jordan M and Dawson, Eric T and Jones, William and Lin, Michael F and Paten, Benedict and Durbin, Richard},
     title = {Variation graph toolkit improves read mapping by representing genetic variation in the reference},
     journal = {Nature Biotechnology},
     volume = {36},
@@ -33,7 +32,7 @@
     doi = {doi:10.1038/nbt.4227}
 }
 @article{Bandage,
-    author = {Wick R.R., Schultz M.B., Zobel J. and Holt K.E.},
+    author = {Wick, R.R. and Schultz, M.B. and Zobel, J. and Holt, K.E.},
     year = {2015},
     title = {Bandage: interactive visualisation of de novo genome assemblies},
     journal = {Bioinformatics},
@@ -48,13 +47,13 @@
     url = {http://lh3.github.io/2014/07/19/a-proposal-of-the-grapical-fragment-assembly-format}
 }
 @online{GFA-SPEC,
-    author = {Gene Myers, Shaun Jackman, Giorgio Gonnella, Jason Chin, and Richard Durbin},
+    author = {Myers, Gene and Jackman, Shaun and Gonnella, Giorgio and Chin, Jason and Durbin, Richard},
     title = {Graphical Fragment Assembly (GFA) Format Specification},
     url = {https://github.com/GFA-spec/GFA-spec},
     year = {2015}
 }
 @article{GfaPy,
-    author = {Giorgio Gonnella and Stefan Kurtz},
+    author = {Gonnella, Giorgio and Kurtz, Stefan},
     title = {GfaPy: a flexible and extensible software library for handling sequence graphs in Python},
     journal = {Bioinformatics},
     volume = {33},
@@ -64,7 +63,7 @@
 
 }
 @article{RGFA,
-    author = {Giorgio Gonnella, Stefan Kurtz},
+    author = {Gonnella, Giorgio and Kurtz, Stefan},
     title = {RGFA: powerful and convenient handling of assembly graphs},
     journal = {PeerJ},
     volume = {4},
[…]
