### 2019-01-22 #1083 @edawson
@mgymrek I had to make a few changes to the bibliography that I missed earlier, but the latest proof is now good. The DOI is: 10.5281/zenodo.2546721 . Thanks (and thanks again @sjackman )!
