+++ Makefile
@@ -34,7 +34,7 @@ $(BIN_DIR):
 
 install: gfak
 	mkdir -p $(DESTDIR)$(PREFIX)/include
-  mkdir -p $(DESTDIR)$(PREFIX)/bin
+	mkdir -p $(DESTDIR)$(PREFIX)/bin
 	cp src/gfakluge.hpp $(DESTDIR)$(PREFIX)/include/
 	cp src/tinyFA/tinyfa.hpp $(DESTDIR)$(PREFIX)/include/
 	cp src/tinyFA/pliib.hpp $(DESTDIR)$(PREFIX)/include/
