### 2019-01-22 #1083 @sjackman
I've added the citation to the Brewsci/bio package. https://github.com/brewsci/homebrew-bio/commit/311543ab3629b3fd029fdd16be59c2040867607b
