### 2019-01-22 #1083 @sjackman
The URL https://doi.org/10.21105/joss.01083 resolves to JOSS, but doi.org does not yet have a Bibtex entry. I imagine it will soon enough.
```console
$ curl -sLH "Accept: text/bibliography; style=bibtex" "https://doi.org/10.21105/joss.01083"
Resource not found.
```
