### 2019-01-22 #1083 @arfon
> The URL https://doi.org/10.21105/joss.01083 resolves to JOSS, but doi.org does not yet have a Bibtex entry. I imagine it will soon enough.

Yeah, it can take a few hours for the DOI metadata to propagate properly.
